+++ behave/main.py
@@ -87,22 +87,25 @@ def main():
         sys.exit(0)
 
     if not config.format:
-        format0 = config.defaults["format0"]
-        config.format = [ format0 ]
+        default_format = config.defaults["default_format"]
+        config.format = [ default_format ]
+    elif config.format and "format" in config.defaults:
+        # -- CASE: Formatter are specified in behave configuration file.
+        #    Check if formatter are provided on command-line, too.
+        if len(config.format) == len(config.defaults["format"]):
+            # -- NO FORMATTER on command-line: Add default formatter.
+            default_format = config.defaults["default_format"]
+            config.format.append(default_format)
     elif 'help' in config.format:
         print "Available formatters:"
         formatters.list_formatters(sys.stdout)
         sys.exit(0)
-    # -- SANITY: Use at most one formatter, more cause various problems.
-    # PROBLEM DESCRIPTION:
-    #   1. APPEND MODE: configfile.format + --format
-    #   2. Daisy chaining of formatters does not work
-    #     => behave.formatter.formatters.get_formatter()
-    #     => Stream methods, stream.write(), stream.flush are missing
-    #        in Formatter interface
-    config.format = config.format[-1:]
 
-    stream = config.output
+    if len(config.outputs) > len(config.format):
+        print 'CONFIG-ERROR: More outfiles (%d) than formatters (%d).' % \
+              (len(config.outputs), len(config.format))
+        sys.exit(1)
+
     runner = Runner(config)
     try:
         failed = runner.run()