+++ pavement.py
@@ -1,7 +1,7 @@
 # ============================================================================
 # PAVER MAKEFILE (pavement.py) -- behave
 # ============================================================================
-# REQUIRES: paver >= 1.1
+# REQUIRES: paver >= 1.2
 # DESCRIPTION:
 #   Provides platform-neutral "Makefile" for simple, project-specific tasks.
 #   AVOID: setup support, because it is currently handled elsewhere.
@@ -51,6 +51,7 @@ options(
     ),
     behave_test=Bunch(
         default_args=[
+            "features/",
             "tools/test-features/",
             "selftest.features/",
             "issue.features/",