+++ test/test_model.py
@@ -10,6 +10,7 @@
 
 from __future__ import with_statement
 
+import re
 import sys
 from mock import Mock, patch
 from nose.tools import *
@@ -17,6 +18,7 @@ import unittest
 from behave import model
 from behave.compat.collections import OrderedDict
 from behave import step_registry
+from behave.configuration import Configuration
 
 class TestFeatureRun(unittest.TestCase):
     def setUp(self):
@@ -24,7 +26,7 @@ class TestFeatureRun(unittest.TestCase):
         self.runner.feature.tags = []
         self.config = self.runner.config = Mock()
         self.context = self.runner.context = Mock()
-        self.formatter = self.runner.formatter = Mock()
+        self.formatters = self.runner.formatters = [Mock()]
         self.run_hook = self.runner.run_hook = Mock()
 
     def test_formatter_feature_called(self):
@@ -33,7 +35,7 @@ class TestFeatureRun(unittest.TestCase):
 
         feature.run(self.runner)
 
-        self.formatter.feature.assert_called_with(feature)
+        self.formatters[0].feature.assert_called_with(feature)
 
     def test_formatter_background_called_when_feature_has_background(self):
         feature = model.Feature('foo.feature', 1, u'Feature', u'foo',
@@ -41,14 +43,14 @@ class TestFeatureRun(unittest.TestCase):
 
         feature.run(self.runner)
 
-        self.formatter.background.assert_called_with(feature.background)
+        self.formatters[0].background.assert_called_with(feature.background)
 
     def test_formatter_background_not_called_when_feature_has_no_background(self):
         feature = model.Feature('foo.feature', 1, u'Feature', u'foo')
 
         feature.run(self.runner)
 
-        assert not self.formatter.background.called
+        assert not self.formatters[0].background.called
 
     def test_run_runs_scenarios(self):
         scenarios = [Mock(), Mock()]
@@ -57,6 +59,7 @@ class TestFeatureRun(unittest.TestCase):
             scenario.run.return_value = False
 
         self.config.tags.check.return_value = True
+        self.config.name = []
 
         feature = model.Feature('foo.feature', 1, u'Feature', u'foo',
                                 scenarios=scenarios)
@@ -65,6 +68,50 @@ class TestFeatureRun(unittest.TestCase):
 
         for scenario in scenarios:
             scenario.run.assert_called_with(self.runner)
+
+    def test_run_runs_named_scenarios(self):
+        scenarios = [Mock(), Mock()]
+        scenarios[0].name = 'first scenario'
+        scenarios[1].name = 'second scenario'
+        scenarios[0].tags = []
+        scenarios[1].tags = []
+
+        for scenario in scenarios:
+            scenario.run.return_value = False
+
+        self.config.tags.check.return_value = True
+        self.config.name = ['first', 'third']
+        self.config.name_re = Configuration.build_name_re(self.config.name)
+
+        feature = model.Feature('foo.feature', 1, u'Feature', u'foo',
+                                scenarios=scenarios)
+
+        feature.run(self.runner)
+
+        scenarios[0].run.assert_called_with(self.runner)
+        assert not scenarios[1].run.called
+
+    def test_run_runs_named_scenarios_with_regexp(self):
+        scenarios = [Mock(), Mock()]
+        scenarios[0].name = 'first scenario'
+        scenarios[1].name = 'second scenario'
+        scenarios[0].tags = []
+        scenarios[1].tags = []
+
+        for scenario in scenarios:
+            scenario.run.return_value = False
+
+        self.config.tags.check.return_value = True
+        self.config.name = ['third .*', 'second .*']
+        self.config.name_re = Configuration.build_name_re(self.config.name)
+
+        feature = model.Feature('foo.feature', 1, u'Feature', u'foo',
+                                scenarios=scenarios)
+
+        feature.run(self.runner)
+
+        assert not scenarios[0].run.called
+        scenarios[1].run.assert_called_with(self.runner)
 
     def test_feature_hooks_not_run_if_feature_not_being_run(self):
         self.config.tags.check.return_value = False
@@ -83,7 +130,7 @@ class TestScenarioRun(unittest.TestCase):
         self.config = self.runner.config = Mock()
         self.config.dry_run = False
         self.context = self.runner.context = Mock()
-        self.formatter = self.runner.formatter = Mock()
+        self.formatters = self.runner.formatters = [Mock()]
         self.run_hook = self.runner.run_hook = Mock()
 
     def test_run_invokes_formatter_scenario_and_steps_correctly(self):
@@ -96,7 +143,7 @@ class TestScenarioRun(unittest.TestCase):
 
         scenario.run(self.runner)
 
-        self.formatter.scenario.assert_called_with(scenario)
+        self.formatters[0].scenario.assert_called_with(scenario)
         [step.run.assert_called_with(self.runner) for step in steps]
 
     if sys.version_info[0] == 3:
@@ -129,6 +176,7 @@ class TestScenarioRun(unittest.TestCase):
         steps[0].run.return_value = False
         steps[1].step_type = "when"
         steps[1].name = "step1"
+
         def step1_function(context):
             pass
         my_step_registry = step_registry.StepRegistry()
@@ -162,7 +210,7 @@ class TestScenarioRun(unittest.TestCase):
         passed_step = Mock()
         undefined_step = Mock()
 
-        steps = [ passed_step, undefined_step ]
+        steps = [passed_step, undefined_step]
         scenario = model.Scenario('foo.feature', 17, u'Scenario', u'foo',
                                   steps=steps)
         passed_step.run.return_value = True
@@ -268,9 +316,10 @@ class TestStepRun(unittest.TestCase):
     def setUp(self):
         self.runner = Mock()
         self.config = self.runner.config = Mock()
+        self.config.outputs = [None]
         self.context = self.runner.context = Mock()
         print ('context is', self.context)
-        self.formatter = self.runner.formatter = Mock()
+        self.formatters = self.runner.formatters = [Mock()]
         self.step_registry = Mock()
         self.stdout_capture = self.runner.stdout_capture = Mock()
         self.stdout_capture.getvalue.return_value = ''
@@ -296,8 +345,8 @@ class TestStepRun(unittest.TestCase):
         with patch('behave.step_registry.registry', self.step_registry):
             assert not step.run(self.runner)
 
-        self.formatter.match.assert_called_with(model.NoMatch())
-        self.formatter.result.assert_called_with(step)
+        self.formatters[0].match.assert_called_with(model.NoMatch())
+        self.formatters[0].result.assert_called_with(step)
 
     def test_run_with_no_match_does_not_touch_formatter_when_quiet(self):
         step = model.Step('foo.feature', 17, u'Given', 'given', u'foo')
@@ -305,8 +354,8 @@ class TestStepRun(unittest.TestCase):
         with patch('behave.step_registry.registry', self.step_registry):
             assert not step.run(self.runner, quiet=True)
 
-        assert not self.formatter.match.called
-        assert not self.formatter.result.called
+        assert not self.formatters[0].match.called
+        assert not self.formatters[0].result.called
 
     def test_run_when_not_quiet_reports_match_and_result(self):
         step = model.Step('foo.feature', 17, u'Given', 'given', u'foo')
@@ -319,8 +368,8 @@ class TestStepRun(unittest.TestCase):
             match.run.side_effect = side_effect
             with patch('behave.step_registry.registry', self.step_registry):
                 step.run(self.runner)
-            self.formatter.match.assert_called_with(match)
-            self.formatter.result.assert_called_with(step)
+            self.formatters[0].match.assert_called_with(match)
+            self.formatters[0].result.assert_called_with(step)
 
     def test_run_when_quiet_reports_nothing(self):
         step = model.Step('foo.feature', 17, u'Given', 'given', u'foo')
@@ -332,8 +381,8 @@ class TestStepRun(unittest.TestCase):
         for side_effect in side_effects:
             match.run.side_effect = side_effect
             step.run(self.runner, quiet=True)
-            assert not self.formatter.match.called
-            assert not self.formatter.result.called
+            assert not self.formatters[0].match.called
+            assert not self.formatters[0].result.called
 
     def test_run_runs_before_hook_then_match_then_after_hook(self):
         __pychecker__ = "unusednames=args,kwargs"
@@ -497,6 +546,7 @@ class TestTableModel(unittest.TestCase):
         [u'lint', u'low', u'high'],
         [u'green', u'variable', u'awkward'],
     ]
+
     def setUp(self):
         self.table = model.Table(self.HEAD, 0, self.DATA)